apps/server/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.7
qdrant-client==1.10.0
pymongo==4.8.0
requests==2.32.3
